muen-ai-service/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pandas
numpy